fix: 패키지 이름 수정 #7
